Properly coerce unit structs to values where applicable
