+++ compiler/hash-typecheck/src/traverse/scopes.rs
@@ -4,7 +4,7 @@
 use hash_ast::{ast, visitor::AstVisitor};
 use hash_source::identifier::Identifier;
 
-use super::TcVisitor;
+use super::visitor::TcVisitor;
 use crate::{
     diagnostics::error::TcResult,
     ops::{scope::ScopeManager, AccessToOps},
