+++ tests/cases/typecheck/enums/adt.hash
@@ -6,6 +6,6 @@ ensure := <T> => (t: T) => {};
 
 main := () => {
   ensure<Vote>(Vote::Left(1usize));
-  ensure<Vote>(Vote::Abstain());
+  ensure<Vote>(Vote::Abstain);
   ensure<(usize) -> Vote>((x: usize) => Vote::Left(x));
 };
