+++ tests/cases/typecheck/enums/basic.hash
@@ -5,14 +5,15 @@ PrimaryColour := enum(Red, Green, Blue);
 ensure := <T> => (t: T) => {};
 
 main := () => {
-  ensure<PrimaryColour>(PrimaryColour::Red());
-  ensure<PrimaryColour>(PrimaryColour::Green());
-  ensure<PrimaryColour>(PrimaryColour::Blue());
+  ensure<PrimaryColour>(PrimaryColour::Red);
+  ensure<PrimaryColour>(PrimaryColour::Green);
+  ensure<PrimaryColour>(PrimaryColour::Blue);
 
-  red := PrimaryColour::Red();
+  red := PrimaryColour::Red;
+  ensure<PrimaryColour::Red>(red);
   ensure<PrimaryColour>(red);
 
-  many := [PrimaryColour::Red() as PrimaryColour, PrimaryColour::Blue() as PrimaryColour, PrimaryColour::Green() as PrimaryColour];
+  many := [PrimaryColour::Red as PrimaryColour, PrimaryColour::Blue as PrimaryColour, PrimaryColour::Green as PrimaryColour];
   ensure<List<PrimaryColour>>(many);
 };
